+++ mooney/src/main/java/tamtam/mooney/domain/controller/ReportController.java
@@ -5,6 +5,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import tamtam.mooney.domain.dto.response.CategoryExpenseResponseDto;
+import tamtam.mooney.domain.dto.response.MonthlyReportResponseDto;
 import tamtam.mooney.domain.service.CategoryExpenseService;
 import tamtam.mooney.domain.service.MonthlyReportService;
 
@@ -20,13 +22,15 @@ public class ReportController {
     @PostMapping("/monthly")
     public ResponseEntity<?> getMonthlyExpenseReport(@RequestParam final int year,
                                                      @RequestParam final int month) {
-        return ResponseEntity.ok(monthlyReportService.getMonthlyReport(year, month));
+        MonthlyReportResponseDto responseDto = monthlyReportService.getMonthlyReport(year, month);
+        return ResponseEntity.ok(responseDto);
     }
 
     @Operation(summary = "월 소비 카테고리 별 퍼센트, 지난 달 반영 예산값 가져오기")
     @PostMapping("/categories")
     public ResponseEntity<?> getCategoryExpenseStats(@RequestParam final int year,
                                                      @RequestParam final int month) {
-        return ResponseEntity.ok(categoryExpenseService.getCategoryExpenseStats(year, month));
+        CategoryExpenseResponseDto responseDto = categoryExpenseService.getCategoryExpenseStats(year, month);
+        return ResponseEntity.ok(responseDto);
     }
 }
