--- mooney/src/main/java/tamtam/mooney/domain/dto/DailyBudgetRequestDto.java
@@ -1,16 +1,0 @@
-package tamtam.mooney.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class DailyBudgetRequestDto {
-    private List<Long> scheduleIds;
-}
