--- mooney/src/main/java/tamtam/mooney/domain/dto/MonthlyBudgetRequestDto.java
@@ -1,9 +1,0 @@
-package tamtam.mooney.domain.dto;
-
-import java.math.BigDecimal;
-
-public record MonthlyBudgetRequestDto (
-        BigDecimal budgetAmount,
-        String message
-) {
-}
