+++ mooney/src/main/java/tamtam/mooney/domain/dto/request/DailyBudgetRequestDto.java
@@ -1,0 +1,8 @@
+package tamtam.mooney.domain.dto.request;
+
+import javax.validation.constraints.NotNull;
+import java.util.List;
+
+public record DailyBudgetRequestDto (
+        @NotNull List<Long> scheduleIds
+){ }
