+++ mooney/src/main/java/tamtam/mooney/domain/dto/request/MonthlyBudgetRequestDto.java
@@ -1,0 +1,10 @@
+package tamtam.mooney.domain.dto.request;
+
+import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
+
+public record MonthlyBudgetRequestDto (
+        @NotNull BigDecimal budgetAmount,
+        @NotNull String message
+) {
+}
