+++ mooney/src/main/java/tamtam/mooney/domain/dto/response/MonthlyReportResponseDto.java
@@ -1,4 +1,4 @@
-package tamtam.mooney.domain.dto;
+package tamtam.mooney.domain.dto.response;
 
 import lombok.Builder;
