+++ mooney/src/main/java/tamtam/mooney/domain/service/MonthlyBudgetService.java
@@ -4,8 +4,8 @@ import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import tamtam.mooney.domain.dto.MonthlyBudgetRequestDto;
-import tamtam.mooney.domain.dto.MonthlyBudgetResponseDto;
+import tamtam.mooney.domain.dto.request.MonthlyBudgetRequestDto;
+import tamtam.mooney.domain.dto.response.MonthlyBudgetResponseDto;
 import tamtam.mooney.domain.dto.GptMonthlyBudgetResponseDto;
 import tamtam.mooney.domain.entity.MonthlyBudget;
 import tamtam.mooney.domain.entity.User;
